// web-ui/client/gen/sales/quote.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */
import type { QuoteLineItem } from './quoteLineItem';

export interface Quote {
  readonly id: number;
  opportunity: number;
  readonly opportunity_name: string;
  /** @maxLength 255 */
  title: string;
  /** @pattern ^-?\d{0,10}(?:\.\d{0,2})?$ */
  total_price?: string;
  /** @nullable */
  created_by?: number | null;
  readonly created_by_name: string;
  readonly created_at: string;
  notes?: string;
  readonly line_items: readonly QuoteLineItem[];
}